# docs/05-api/openapi.yaml
openapi: 3.1.0
info:
  title: Mira API
  version: 1.0.0
  description: |
    Mira — your closet, your stylist, your mirror.

    Canonical contract for the Mira API. Enumerated values come from
    docs/04-data/taxonomy.md and MUST NOT be extended by implementations.

    Rules that apply to every operation:
      - Responses contain only the authenticated user's data.
      - A resource owned by another user returns 404, never 403.
      - Every creating POST requires an Idempotency-Key header.
      - Image URLs are signed and expiring; clients must honour urlExpiresAt.
servers:
  - url: https://api.mira.app/v1
    description: Production
  - url: https://api.staging.mira.app/v1
    description: Staging
  - url: http://localhost:4000/v1
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /imports/photo:
    post:
      tags: [ingestion]
      summary: Import a garment from an uploaded photo
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [upload_key]
              properties:
                upload_key: { type: string }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GarmentAccepted'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, maxLength: 128 }
...
  schemas:
...
    GarmentAccepted:
      type: object
      properties:
        garment_id: { type: string, format: uuid }
        job_id: { type: string, format: uuid }
...